app/ecg_core/__init__.py

import os 
os.environ["CUDA_VISIBLE_DEVICES"] = "-1"

import pywt
import numpy as np
import pandas as pd
import padasip as pa 
from scipy import sparse
from scipy.sparse.linalg import spsolve
from scipy.signal import find_peaks 
from scipy.signal import resample, resample_poly
from sklearn.preprocessing import MinMaxScaler, MaxAbsScaler
from datetime import timedelta, datetime
from ecgdetectors import Detectors
from keras.models import load_model