requirements.txt
Flask
Flask-Admin
Flask-SQLAlchemy
Flask-Security>=1.7.5
enum34==1.1.6
python-socketio==4.6.0
flask-socketio==4.3.1
flask_sqlalchemy
flask_wtf
wtforms
wtforms[email]

greenlet
eventlet
gevent
gevent-websocket

numpy
matplotlib
pandas
padasip
scipy
PyWavelets
scikit-learn
py-ecg-detectors
pickle-mixin

tensorflow-cpu==2.1.0
keras==2.3.1